Move user helper methods to new module
We want to re-use these methods to get the options for filters in the
users index view. Move them to a separarte helper module to be
included in the FilterInput without having to import ApplicationHelper.

Change calls to `t` to `I18n.t` so that these methods can be called
from non-view contexts.
